Add files via upload
The "Sticky Notes" web application allows users to create and manage notes. Upon accessing the site, users can add new notes, which are displayed in a list. If no notes are present, the interface prompts users to add their first note. The application provides a simple and intuitive interface for note-taking and organization. 

https://sticky-notes-e7a99.web.app/